acme_medical/requirements.txt
Django==1.7
PyYAML==3.11
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==2.4.3
gunicorn==19.1.1
heroku==0.1.4
psycopg2==2.5.4
python-dateutil==1.5
requests==2.4.3
static3==0.5.1
wsgiref==0.1.2
